refactor

+++ src/main/java/net/ccbluex/liquidbounce/launch/data/legacyui/clickgui/style/styles/NullStyle.java
@@ -35,8 +35,8 @@ public class NullStyle extends Style {
 
     @Override
     public void drawPanel(int mouseX, int mouseY, Panel panel) {
-        RenderUtils.drawGradientSidewaysH((float) panel.getX() - 3, (float) panel.getY(), (float) panel.getX() + panel.getWidth() + 3, (float) panel.getY() + 19, ClickGUIModule.generateColor().getRGB(), ColorUtils.INSTANCE.reAlpha(ClickGUIModule.generateColor(),50).getRGB());
-        if(panel.getFade() > 0)
+        RenderUtils.drawGradientSidewaysH((float) panel.getX() - 3, (float) panel.getY(), (float) panel.getX() + panel.getWidth() + 3, (float) panel.getY() + 19, ClickGUIModule.generateColor().getRGB(), ColorUtils.INSTANCE.reAlpha(ClickGUIModule.generateColor(), 50).getRGB());
+        if (panel.getFade() > 0)
             RenderUtils.drawBorderedRect((float) panel.getX(), (float) panel.getY() + 19, (float) panel.getX() + panel.getWidth(), panel.getY() + 19 + panel.getFade(), 1, Integer.MIN_VALUE, Integer.MIN_VALUE);
         GlStateManager.resetColor();
         float textWidth = Fonts.minecraftFont.getStringWidth("§f" + StringUtils.stripControlCodes(panel.getName()));
@@ -62,30 +62,30 @@ public class NullStyle extends Style {
     public void drawModuleElement(int mouseX, int mouseY, ModuleElement moduleElement) {
         final int guiColor = ClickGUIModule.generateColor().getRGB();
         GlStateManager.resetColor();
-        Fonts.minecraftFont.drawString(moduleElement.getDisplayName(), (int) (moduleElement.getX()+3), moduleElement.getY() + 7, moduleElement.getModule().getState() ? guiColor : Integer.MAX_VALUE);
+        Fonts.minecraftFont.drawString(moduleElement.getDisplayName(), moduleElement.getX() + 3, moduleElement.getY() + 7, moduleElement.getModule().getState() ? guiColor : Integer.MAX_VALUE);
 
         final List<Value<?>> moduleValues = moduleElement.getModule().getValues();
 
-        if(!moduleValues.isEmpty()) {
+        if (!moduleValues.isEmpty()) {
             Fonts.minecraftFont.drawString("+", moduleElement.getX() + moduleElement.getWidth() - 8, moduleElement.getY() + (moduleElement.getHeight() / 2), Color.WHITE.getRGB());
 
-            if(moduleElement.isShowSettings()) {
+            if (moduleElement.isShowSettings()) {
                 int yPos = moduleElement.getY() + 4;
-                for(final Value value : moduleValues) {
-                    if(!value.getDisplayable())
+                for (final Value value : moduleValues) {
+                    if (!value.getDisplayable())
                         continue;
 
-                    if(value instanceof BoolValue) {
+                    if (value instanceof BoolValue) {
                         String text = value.getName();
                         float textWidth = Fonts.minecraftFont.getStringWidth(text);
 
-                        if(moduleElement.getSettingsWidth() < textWidth + 8)
+                        if (moduleElement.getSettingsWidth() < textWidth + 8)
                             moduleElement.setSettingsWidth(textWidth + 8);
 
                         RenderUtils.drawRect(moduleElement.getX() + moduleElement.getWidth() + 4, yPos + 2, moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth(), yPos + 14, Integer.MIN_VALUE);
 
-                        if(mouseX >= moduleElement.getX() + moduleElement.getWidth() + 4 && mouseX <= moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth() && mouseY >= yPos + 2 && mouseY <= yPos + 14) {
-                            if(Mouse.isButtonDown(0) && moduleElement.isntPressed()) {
+                        if (mouseX >= moduleElement.getX() + moduleElement.getWidth() + 4 && mouseX <= moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth() && mouseY >= yPos + 2 && mouseY <= yPos + 14) {
+                            if (Mouse.isButtonDown(0) && moduleElement.isntPressed()) {
                                 final BoolValue boolValue = (BoolValue) value;
 
                                 boolValue.set(!boolValue.get());
@@ -96,13 +96,13 @@ public class NullStyle extends Style {
                         GlStateManager.resetColor();
                         Fonts.minecraftFont.drawString(text, moduleElement.getX() + moduleElement.getWidth() + 6, yPos + 4, ((BoolValue) value).get() ? guiColor : Integer.MAX_VALUE);
                         yPos += 12;
-                    }else if(value instanceof ListValue) {
+                    } else if (value instanceof ListValue) {
                         ListValue listValue = (ListValue) value;
 
                         String text = value.getName();
                         float textWidth = Fonts.minecraftFont.getStringWidth(text);
 
-                        if(moduleElement.getSettingsWidth() < textWidth + 16)
+                        if (moduleElement.getSettingsWidth() < textWidth + 16)
                             moduleElement.setSettingsWidth(textWidth + 16);
 
                         RenderUtils.drawRect(moduleElement.getX() + moduleElement.getWidth() + 4, yPos + 2, moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth(), yPos + 14, Integer.MIN_VALUE);
@@ -110,8 +110,8 @@ public class NullStyle extends Style {
                         Fonts.minecraftFont.drawString("§c" + text, moduleElement.getX() + moduleElement.getWidth() + 6, yPos + 4, 0xffffff);
                         Fonts.minecraftFont.drawString(listValue.openList ? "-" : "+", (int) (moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth() - (listValue.openList ? 5 : 6)), yPos + 4, 0xffffff);
 
-                        if(mouseX >= moduleElement.getX() + moduleElement.getWidth() + 4 && mouseX <= moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth() && mouseY >= yPos + 2 && mouseY <= yPos + 14) {
-                            if(Mouse.isButtonDown(0) && moduleElement.isntPressed()) {
+                        if (mouseX >= moduleElement.getX() + moduleElement.getWidth() + 4 && mouseX <= moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth() && mouseY >= yPos + 2 && mouseY <= yPos + 14) {
+                            if (Mouse.isButtonDown(0) && moduleElement.isntPressed()) {
                                 listValue.openList = !listValue.openList;
                                 mc.getSoundHandler().playSound(PositionedSoundRecord.create(new ResourceLocation("gui.button.press"), 1.0F));
                             }
@@ -119,17 +119,17 @@ public class NullStyle extends Style {
 
                         yPos += 12;
 
-                        for(final String valueOfList : listValue.getValues()) {
+                        for (final String valueOfList : listValue.getValues()) {
                             final float textWidth2 = Fonts.minecraftFont.getStringWidth(">" + valueOfList);
 
-                            if(moduleElement.getSettingsWidth() < textWidth2 + 12)
+                            if (moduleElement.getSettingsWidth() < textWidth2 + 12)
                                 moduleElement.setSettingsWidth(textWidth2 + 12);
 
                             if (listValue.openList) {
                                 RenderUtils.drawRect(moduleElement.getX() + moduleElement.getWidth() + 4, yPos + 2, moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth(), yPos + 14, Integer.MIN_VALUE);
 
-                                if(mouseX >= moduleElement.getX() + moduleElement.getWidth() + 4 && mouseX <= moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth() && mouseY >= yPos + 2 && mouseY <= yPos + 14) {
-                                    if(Mouse.isButtonDown(0) && moduleElement.isntPressed()) {
+                                if (mouseX >= moduleElement.getX() + moduleElement.getWidth() + 4 && mouseX <= moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth() && mouseY >= yPos + 2 && mouseY <= yPos + 14) {
+                                    if (Mouse.isButtonDown(0) && moduleElement.isntPressed()) {
                                         listValue.set(valueOfList);
                                         mc.getSoundHandler().playSound(PositionedSoundRecord.create(new ResourceLocation("gui.button.press"), 1.0F));
                                     }
@@ -141,12 +141,12 @@ public class NullStyle extends Style {
                                 yPos += 12;
                             }
                         }
-                    }else if(value instanceof FloatValue) {
+                    } else if (value instanceof FloatValue) {
                         FloatValue floatValue = (FloatValue) value;
                         String text = value.getName() + "§f: §c" + round(floatValue.get());
                         float textWidth = Fonts.minecraftFont.getStringWidth(text);
 
-                        if(moduleElement.getSettingsWidth() < textWidth + 8)
+                        if (moduleElement.getSettingsWidth() < textWidth + 8)
                             moduleElement.setSettingsWidth(textWidth + 8);
 
                         RenderUtils.drawRect(moduleElement.getX() + moduleElement.getWidth() + 4, yPos + 2, moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth(), yPos + 24, Integer.MIN_VALUE);
@@ -154,8 +154,8 @@ public class NullStyle extends Style {
                         float sliderValue = moduleElement.getX() + moduleElement.getWidth() + ((moduleElement.getSettingsWidth() - 12) * (floatValue.get() - floatValue.getMinimum()) / (floatValue.getMaximum() - floatValue.getMinimum()));
                         RenderUtils.drawRect(8 + sliderValue, yPos + 15, sliderValue + 11, yPos + 21, guiColor);
 
-                        if(mouseX >= moduleElement.getX() + moduleElement.getWidth() + 4 && mouseX <= moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth() - 4 && mouseY >= yPos + 15 && mouseY <= yPos + 21) {
-                            if(Mouse.isButtonDown(0)) {
+                        if (mouseX >= moduleElement.getX() + moduleElement.getWidth() + 4 && mouseX <= moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth() - 4 && mouseY >= yPos + 15 && mouseY <= yPos + 21) {
+                            if (Mouse.isButtonDown(0)) {
                                 double i = MathHelper.clamp_double((mouseX - moduleElement.getX() - moduleElement.getWidth() - 8) / (moduleElement.getSettingsWidth() - 12), 0, 1);
                                 floatValue.set(round((float) (floatValue.getMinimum() + (floatValue.getMaximum() - floatValue.getMinimum()) * i)).floatValue());
                             }
@@ -164,12 +164,12 @@ public class NullStyle extends Style {
                         GlStateManager.resetColor();
                         Fonts.minecraftFont.drawString(text, moduleElement.getX() + moduleElement.getWidth() + 6, yPos + 4, 0xffffff);
                         yPos += 22;
-                    }else if(value instanceof IntegerValue) {
+                    } else if (value instanceof IntegerValue) {
                         IntegerValue integerValue = (IntegerValue) value;
                         String text = value.getName() + "§f: §c" + (value instanceof BlockValue ? BlockUtils.getBlockName(integerValue.get()) + " (" + integerValue.get() + ")" : integerValue.get());
                         float textWidth = Fonts.minecraftFont.getStringWidth(text);
 
-                        if(moduleElement.getSettingsWidth() < textWidth + 8)
+                        if (moduleElement.getSettingsWidth() < textWidth + 8)
                             moduleElement.setSettingsWidth(textWidth + 8);
 
                         RenderUtils.drawRect(moduleElement.getX() + moduleElement.getWidth() + 4, yPos + 2, moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth(), yPos + 24, Integer.MIN_VALUE);
@@ -177,8 +177,8 @@ public class NullStyle extends Style {
                         float sliderValue = moduleElement.getX() + moduleElement.getWidth() + ((moduleElement.getSettingsWidth() - 12) * (integerValue.get() - integerValue.getMinimum()) / (integerValue.getMaximum() - integerValue.getMinimum()));
                         RenderUtils.drawRect(8 + sliderValue, yPos + 15, sliderValue + 11, yPos + 21, guiColor);
 
-                        if(mouseX >= moduleElement.getX() + moduleElement.getWidth() + 4 && mouseX <= moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth() && mouseY >= yPos + 15 && mouseY <= yPos + 21) {
-                            if(Mouse.isButtonDown(0)) {
+                        if (mouseX >= moduleElement.getX() + moduleElement.getWidth() + 4 && mouseX <= moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth() && mouseY >= yPos + 15 && mouseY <= yPos + 21) {
+                            if (Mouse.isButtonDown(0)) {
                                 double i = MathHelper.clamp_double((mouseX - moduleElement.getX() - moduleElement.getWidth() - 8) / (moduleElement.getSettingsWidth() - 12), 0, 1);
                                 integerValue.set((int) (integerValue.getMinimum() + (integerValue.getMaximum() - integerValue.getMinimum()) * i));
                             }
@@ -187,7 +187,7 @@ public class NullStyle extends Style {
                         GlStateManager.resetColor();
                         Fonts.minecraftFont.drawString(text, moduleElement.getX() + moduleElement.getWidth() + 6, yPos + 4, 0xffffff);
                         yPos += 22;
-                    }else if(value instanceof FontValue) {
+                    } else if (value instanceof FontValue) {
                         final FontValue fontValue = (FontValue) value;
                         final FontRenderer fontRenderer = fontValue.get();
 
@@ -199,12 +199,12 @@ public class NullStyle extends Style {
                             final GameFontRenderer liquidFontRenderer = (GameFontRenderer) fontRenderer;
 
                             displayString = "Font: " + liquidFontRenderer.getDefaultFont().getFont().getName() + " - " + liquidFontRenderer.getDefaultFont().getFont().getSize();
-                        }else if(fontRenderer == Fonts.minecraftFont)
+                        } else if (fontRenderer == Fonts.minecraftFont)
                             displayString = "Font: Minecraft";
-                        else{
+                        else {
                             final Object[] objects = Fonts.getFontDetails(fontRenderer);
 
-                            if(objects != null) {
+                            if (objects != null) {
                                 displayString = objects[0] + ((int) objects[1] != -1 ? " - " + objects[1] : "");
                             }
                         }
@@ -212,37 +212,37 @@ public class NullStyle extends Style {
                         Fonts.minecraftFont.drawString(displayString, moduleElement.getX() + moduleElement.getWidth() + 6, yPos + 4, Color.WHITE.getRGB());
                         int stringWidth = Fonts.minecraftFont.getStringWidth(displayString);
 
-                        if(moduleElement.getSettingsWidth() < stringWidth + 8)
+                        if (moduleElement.getSettingsWidth() < stringWidth + 8)
                             moduleElement.setSettingsWidth(stringWidth + 8);
 
-                        if((Mouse.isButtonDown(0) && !mouseDown || Mouse.isButtonDown(1) && !rightMouseDown) && mouseX >= moduleElement.getX() + moduleElement.getWidth() + 4 && mouseX <= moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth() && mouseY >= yPos + 4 && mouseY <= yPos + 12) {
+                        if ((Mouse.isButtonDown(0) && !mouseDown || Mouse.isButtonDown(1) && !rightMouseDown) && mouseX >= moduleElement.getX() + moduleElement.getWidth() + 4 && mouseX <= moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth() && mouseY >= yPos + 4 && mouseY <= yPos + 12) {
                             final List<FontRenderer> fonts = Fonts.getFonts();
 
-                            if(Mouse.isButtonDown(0)) {
-                                for(int i = 0; i < fonts.size(); i++) {
+                            if (Mouse.isButtonDown(0)) {
+                                for (int i = 0; i < fonts.size(); i++) {
                                     final FontRenderer font = fonts.get(i);
 
-                                    if(font == fontRenderer) {
+                                    if (font == fontRenderer) {
                                         i++;
 
-                                        if(i >= fonts.size())
+                                        if (i >= fonts.size())
                                             i = 0;
 
                                         fontValue.set(fonts.get(i));
                                         break;
                                     }
                                 }
-                            }else{
-                                for(int i = fonts.size() - 1; i >= 0; i--) {
+                            } else {
+                                for (int i = fonts.size() - 1; i >= 0; i--) {
                                     final FontRenderer font = fonts.get(i);
 
-                                    if(font == fontRenderer) {
+                                    if (font == fontRenderer) {
                                         i--;
 
-                                        if(i >= fonts.size())
+                                        if (i >= fonts.size())
                                             i = 0;
 
-                                        if(i < 0)
+                                        if (i < 0)
                                             i = fonts.size() - 1;
 
                                         fontValue.set(fonts.get(i));
@@ -253,11 +253,11 @@ public class NullStyle extends Style {
                         }
 
                         yPos += 11;
-                    }else{
+                    } else {
                         String text = value.getName() + "§f: §c" + value.get();
                         float textWidth = Fonts.minecraftFont.getStringWidth(text);
 
-                        if(moduleElement.getSettingsWidth() < textWidth + 8)
+                        if (moduleElement.getSettingsWidth() < textWidth + 8)
                             moduleElement.setSettingsWidth(textWidth + 8);
 
                         RenderUtils.drawRect(moduleElement.getX() + moduleElement.getWidth() + 4, yPos + 2, moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth(), yPos + 14, Integer.MIN_VALUE);
@@ -271,7 +271,7 @@ public class NullStyle extends Style {
                 mouseDown = Mouse.isButtonDown(0);
                 rightMouseDown = Mouse.isButtonDown(1);
 
-                if(moduleElement.getSettingsWidth() > 0F && yPos > moduleElement.getY() + 4)
+                if (moduleElement.getSettingsWidth() > 0F && yPos > moduleElement.getY() + 4)
                     RenderUtils.drawBorderedRect(moduleElement.getX() + moduleElement.getWidth() + 4, moduleElement.getY() + 6, moduleElement.getX() + moduleElement.getWidth() + moduleElement.getSettingsWidth(), yPos + 2, 1F, Integer.MIN_VALUE, 0);
             }
         }
